moving stats into topology and removing float64 support of standardDeviationList
